change: modified the invoice route
